Add additional primitives and parsing

+++ geometry/src/main/java/less/lgeo/primitive/Line.java
@@ -5,14 +5,14 @@ public class Line {
   private static final LineType type = LineType.LINE;
 
   private final Color color;
-  private final int x1;
-  private final int y1;
-  private final int z1;
-  private final int x2;
-  private final int y2;
-  private final int z2;
+  private final double x1;
+  private final double y1;
+  private final double z1;
+  private final double x2;
+  private final double y2;
+  private final double z2;
 
-  public Line(Color color, int x1, int y1, int z1, int x2, int y2, int z2) {
+  public Line(Color color, double x1, double y1, double z1, double x2, double y2, double z2) {
     this.color = color;
     this.x1 = x1;
     this.y1 = y1;
