+++ geometry/src/main/java/less/lgeo/primitive/SubFileReference.java
@@ -1,6 +1,6 @@
 package less.lgeo.primitive;
 
-import less.lgeo.geometery.Matrix;
+import less.lgeo.Matrix;
 
 public class SubFileReference {
     private static final LineType type = LineType.SUB_FILE_REF;
@@ -9,7 +9,8 @@ public class SubFileReference {
     private final Matrix matrix;
     private final String fileReference;
 
-    public SubFileReference(Color color, int x, int y, int z, int a, int b, int c, int d, int e, int f, int g, int h, int i, String fileReference) {
+    public SubFileReference(Color color, double x, double y, double z, double a, double b, double c, double d, double e,
+            double f, double g, double h, double i, String fileReference) {
         this.color = color;
         this.matrix = new Matrix(x, y, z, a, b, c, d, e, f, g, h, i);
         this.fileReference = fileReference;
@@ -24,15 +25,13 @@ public class SubFileReference {
     public Color getColor() {
         return this.color;
     }
-    public Matrix getMatrix()
-    {
+
+    public Matrix getMatrix() {
         return this.matrix;
     }
 
-    public String getFileReference()
-    {
+    public String getFileReference() {
         return this.fileReference;
     }
-
 
 }
